Start adding the missing functions to lapi.rs. (#4)
* Start adding the missing functions to lapi.rs.

These seem to just be used from the rust side, with the C version not being deleted
(and no #[no_mangle]) so I've done the same for the functions that I've added here
and just removed the C externs from the rust code.

Is this the right thing to do? I.e. how do I decide what to just copy to rust and
what to actually wholesale move and delete from the C side?

* cargo fmt

* Remove moved stuff

* Remove transpiled file

* Cleanups

* cargo fmt